requirements.txt
asgiref==3.4.1
Django==3.2.9
django-js-asset==1.2.2
django-mptt==0.13.4
Pillow==8.4.0
pytz==2021.3
shortuuid==1.0.8
sqlparse==0.4.2